% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getIntensityMatrix}
\alias{getIntensityMatrix}
\title{Create intensity matrix from the set of files}
\usage{
getIntensityMatrix(files, path, round = FALSE, digits = 0,
  snrT = 1.5)
}
\arguments{
\item{files}{list of file names to be combined into matrix}

\item{path}{path to the directory where files are located}

\item{round}{where to round the MZ values and aggregate intensities}

\item{digits}{number of digits to round mz to}

\item{snrT}{minimal Signal-to-Noise ratio for peak to be included in the output.}
}
\value{
matrix of peaks intensity.
}
\description{
Create intensity matrix from the set of files
}
\details{
We assume that files were created by the query from this package and contains
following columns:
* id
* scan
* spectrumid
* diagnosis
* num
* rt
* tic
* mz
* intensity
* norm2tic
* snr
}
