% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getIntensityMatrixSQL}
\alias{getIntensityMatrixSQL}
\title{Create intensity matrix from the database}
\usage{
getIntensityMatrixSQL(con, sql, round = FALSE, digits = 0)
}
\arguments{
\item{con}{connection to the database}

\item{sql}{query to take peaks}

\item{round}{where to round the MZ values and aggregate intensities}

\item{digits}{number of digits to round mz to}
}
\value{
matrix of peaks intensity.
}
\description{
Create intensity matrix from the database
}
\details{
SQL query should return three columns in any order: mz, scan, intensity. 
First two will be columns and rows of the matrix, the last one will be matrix 
content. MZ is truncated with round function before matrix generation, 
however it is better to use SQL functions to aggregate intensities of 
appropriate mass, as meaning of aggregation would depend upot type of the 
intensity (normalized, squared etc).
}
