% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{getPeakList}
\alias{getPeakList}
\title{Create peak list from the set of files}
\usage{
getPeakList(files, path, round = FALSE, digits = 0, snrT = 1.5,
  align = FALSE)
}
\arguments{
\item{files}{list of file names to be combined into matrix}

\item{path}{path to the directory where files are located}

\item{round}{where to round the MZ values and aggregate intensities}

\item{digits}{number of digits to round mz to}

\item{snrT}{minimal Signal-to-Noise ratio for peak to be included in the output.}

\item{align}{should result peaklist be aligned}
}
\value{
matrix of peaks intensity.
}
\description{
Create peak list from the set of files
}
\details{
We assume that files were created by the query from this package and contains
following columns:
* id
* scan
* spectrumid
* diagnosis
* num
* rt
* tic
* mz
* intensity
* norm2tic
* snr
}
