% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{makeMassPeak}
\alias{makeMassPeak}
\title{makeMassPeak.}
\usage{
makeMassPeak(peakDT, metadata = NULL, align = FALSE, min_peaks = 10)
}
\arguments{
\item{peakDT}{data.table to take data from}

\item{metadata}{dataFrame with scan-specific metadata.}

\item{align}{logical, should peaks in peakList be aligned.}
}
\value{
list of MassPeaks objects corresponding th data in peakDT
}
\description{
Function converts data.table with columns 'scan','mz' and 'intensity' into list of MassPeaks objects.
}
\details{
Here we assume that the data is homogenious in experimental and measurement
properties such as mode, resolution, diagnosis etc. That assumption is required
as the final step of the conversion is alignment of peak lists which has no
sense for unhomogenious lists. To create unhomogenious peak list create peak lists
for each group, for example diagnosis, and merge them afterwords.
}
