% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{prepareMatrix}
\alias{prepareMatrix}
\title{Prepare Feature Matrix from peak list.}
\usage{
prepareMatrix(pl, gropId, tolerance = 0.002, minFrequency = 0.25,
  mergeWhitelists = TRUE, digits = 3)
}
\arguments{
\item{pl}{list of MassPeaks to convert}

\item{gropId}{group identity for peaklist}

\item{tolerance}{binPeaks tolerance}

\item{minFrequency}{minimal peak frequency for the filterPeaks}

\item{mergeWhitelists}{wether to merge white lists for the filterPeaks}

\item{digits}{number of digits in the Feature matrix column names}
}
\value{
matrix of peaks intensity.
}
\description{
Prepare Feature Matrix from peak list.
}
