% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix.R
\name{makeMassPeak}
\alias{makeMassPeak}
\title{makeMassPeak.}
\usage{
makeMassPeak(peakDT, metadata = NULL)
}
\arguments{
\item{peakDT}{data.table to take data from}

\item{metadata}{dataFrame with scan-specific metadata.}
}
\value{
list of MassPeaks objects corresponding th data in peakDT
}
\description{
Function converts data.table with columns 'scan','mz' and 'intensity' into list of MassPeaks objects.
}
